+++ libs/cardano-protocol/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### `testlib`
 
-* Add `testlib` with `Test.Cardano.Protocol.Arbitrary`, `Test.Cardano.Protocol.Praos.Arbitrary` and `Test.Cardano.Protocol.Leios.Arbitrary`, providing `Arbitrary` instances for `OCert`, `KESPeriod`, `PrevHash`, `InputVRF`, the TPraos `BHeader`/`BHBody`/`Block`, the Praos and Leios `Header`/`HeaderBody`/`Block` and `EbAnnouncement`, and non-annotator `DecCBOR` instances for the Praos and Leios `Header`
+* Add `testlib` with `Test.Cardano.Protocol.TPraos.BlockHeader.Arbitrary`, `Test.Cardano.Protocol.Praos.BlockHeader.Arbitrary` and `Test.Cardano.Protocol.Leios.BlockHeader.Arbitrary`, providing `Arbitrary` instances for `OCert`, `KESPeriod`, `PrevHash`, `InputVRF`, the TPraos `BHeader`/`BHBody`/`Block`, the Praos and Leios `Header`/`HeaderBody`/`Block` and `EbAnnouncement`, and non-annotator `DecCBOR` instances for the TPraos `BHeader` and the Praos and Leios `Header`
 
 ## 0.1.0.0