+++ libs/cardano-protocol/CHANGELOG.md
@@ -2,11 +2,14 @@
 
 ## 0.1.1.0
 
+* Add `LeiosEraBlockHeader` typeclass
+* Add `bodyHeaderL`, `sigHeaderL`
+* Rename `EbAnnouncement` to `EbReferencesAnnouncement`
 * Export `HeaderConstr` from `Cardano.Protocol.Praos.BlockHeader` and `Cardano.Protocol.Leios.BlockHeader`
 
 ### `testlib`
 
-* Add `testlib` with `Test.Cardano.Protocol.TPraos.BlockHeader.Arbitrary`, `Test.Cardano.Protocol.Praos.BlockHeader.Arbitrary` and `Test.Cardano.Protocol.Leios.BlockHeader.Arbitrary`, providing `Arbitrary` instances for `OCert`, `KESPeriod`, `PrevHash`, `InputVRF`, the TPraos `BHeader`/`BHBody`/`Block`, the Praos and Leios `Header`/`HeaderBody`/`Block` and `EbAnnouncement`, and non-annotator `DecCBOR` instances for the TPraos `BHeader` and the Praos and Leios `Header`
+* Add `testlib` with `Test.Cardano.Protocol.TPraos.BlockHeader.Arbitrary`, `Test.Cardano.Protocol.Praos.BlockHeader.Arbitrary` and `Test.Cardano.Protocol.Leios.BlockHeader.Arbitrary`, providing `Arbitrary` instances for `OCert`, `KESPeriod`, `PrevHash`, `InputVRF`, the TPraos `BHeader`/`BHBody`/`Block`, the Praos and Leios `Header`/`HeaderBody`/`Block` and `EbReferencesAnnouncement`, and non-annotator `DecCBOR` instances for the TPraos `BHeader` and the Praos and Leios `Header`
 
 ## 0.1.0.0